pyproject.toml
[project]
name = "kaboudan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arch>=7.2.0",
    "matplotlib>=3.10.0",
    "numpy<2.0",
    "pandas>=2.2.3",
    "pmdarima>=2.0.4",
    "python-bcb>=0.3.0",
    "scikit-learn>=1.6.1",
    "sktime>=0.36.0",
    "yfinance>=0.2.52",
]
